+++ Model/PaypalCommonApi.php
@@ -113,6 +113,10 @@ abstract class PaypalCommonApi
     {
         /** @var Quote $quote */
         $quote = $this->cart->getQuote();
+        $baseSubtotal = $this->cartSalesModelQuote->getBaseSubtotal();
+        $customerBalance = $quote->getData('customer_balance_amount_used');
+        $rewardPoints = $quote->getData('base_reward_currency_amount');
+        $giftCard = $quote->getData('base_gift_cards_amount_used');
 
         /** @var string $storeCurrency */
         $storeCurrency = $quote->getBaseCurrencyCode();
@@ -138,6 +142,48 @@ abstract class PaypalCommonApi
                 ->setDescription('Discount')
                 ->setQuantity('1')
                 ->setPrice($this->cartSalesModelQuote->getBaseDiscountAmount())
+                ->setSku('discountloja')
+                ->setCurrency($storeCurrency);
+            $itemList->addItem($item);
+        }
+
+        if ($customerBalance && $customerBalance !== '0.0000') {
+            if ((float)$customerBalance > 0.0000) {
+                $customerBalance = $customerBalance * (-1);
+            }
+            $item = new Item();
+            $item->setName('Store Credit Discount')
+                ->setDescription('Store Credit Discount')
+                ->setQuantity('1')
+                ->setPrice($customerBalance)
+                ->setSku('discountloja')
+                ->setCurrency($storeCurrency);
+            $itemList->addItem($item);
+        }
+
+        if ($rewardPoints && $rewardPoints !== '0.0000') {
+            if ((float)$rewardPoints > 0.0000) {
+                $rewardPoints = $rewardPoints * (-1);
+            }
+            $item = new Item();
+            $item->setName('Rewards Discount')
+                ->setDescription('Rewards Discount')
+                ->setQuantity('1')
+                ->setPrice($rewardPoints)
+                ->setSku('discountloja')
+                ->setCurrency($storeCurrency);
+            $itemList->addItem($item);
+        }
+
+        if ($giftCard && $giftCard !== '0.0000') {
+            if ((float)$giftCard > 0.0000) {
+                $giftCard = $giftCard * (-1);
+            }
+            $item = new Item();
+            $item->setName('Gift Card Discount')
+                ->setDescription('Gift Card Discount')
+                ->setQuantity('1')
+                ->setPrice($giftCard)
                 ->setSku('discountloja')
                 ->setCurrency($storeCurrency);
             $itemList->addItem($item);
@@ -249,6 +295,17 @@ abstract class PaypalCommonApi
          * then a disscount might be applying, get subtotal with disscount then.
          */
         $baseSubtotal = $this->cartSalesModelQuote->getBaseSubtotal();
+
+        if ($quote->getBaseCustomerBalAmountUsed()) {
+            $baseSubtotal -= $quote->getBaseCustomerBalAmountUsed();
+        }
+        if ($quote->getData('base_reward_currency_amount')) {
+            $baseSubtotal -= $quote->getData('base_reward_currency_amount');
+        }
+
+        if ($quote->getData('base_gift_cards_amount_used')) {
+            $baseSubtotal -= $quote->getData('base_gift_cards_amount_used');
+        }
 
         if ($this->cartSalesModelQuote->getBaseDiscountAmount()) {
             $subtotal = $baseSubtotal + $this->cartSalesModelQuote->getBaseDiscountAmount();